+++ src/test/resources/diff/Foo1.java
@@ -2,10 +2,9 @@ package diff;
 
 class Foo {
 
-    public static void foo() {
-        var x = 1;
-        var y = 2;
-        Bar.bar(x, y);
+    public static void foo(int a) {
+        var b = 1;
+        Bar.bar(a, b);
     }
 
 }