+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.12] - 2022-03-24
+
+### Fixed
+
+- Do not write a data-flow cache file if no results were saved.
+- Data flow engine context regenerated with latest cache after each query.
+
 ## [1.1.11] - 2022-03-24
 
 ### Fixed